requirements.txt
annotated-types==0.5.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
cssselect2==0.7.0
dj-database-url==2.0.0
Django==4.2.2
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap5==23.3
django-heroku==0.3.1
django-pandas==0.6.6
django-pivot==1.9.0
fonttools==4.40.0
gunicorn==20.1.0
html2pdf==0.1
html5lib==1.1
idna==3.4
lxml==4.9.2
mysqlclient==2.1.1
numpy==1.25.2
oscrypto==1.3.0
pandas==2.0.3
pdfkit==1.0.0
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
pydantic==2.0.1
pydantic_core==2.0.2
pydyf==0.6.0
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pypdf==3.9.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.3
weasyprint==59.0
webencodings==0.5.1
whitenoise==6.5.0
xhtml2pdf==0.2.11
zopfli==0.2.2
